ex1 done after multiple tries and with Bryce's help
