+++ src/components/data.js
@@ -1,0 +1,9 @@
+import React from 'react';
+
+const items = [
+    { id: "cursor", name: "Cursor", cost: 10, value: 1 },
+    { id: "grandma", name: "Grandma", cost: 100, value: 10 },
+    { id: "farm", name: "Farm", cost: 1000, value: 80 },
+  ];
+
+  export default items;
